自動回復追加、アイテムウィンドウ変更、バグ修正

+++ Assets/Script/Model/Character/CharacterParams.cs
@@ -31,7 +31,7 @@ namespace Assets.Script.Model
     public class PlayerParams : CharacterParams
     {
         public int Level { get; set; }
-        public int Stamina { get; set; }
+        public float Stamina { get; set; }
         public int MaxStamina { get; set; }
         public override int Attack { get { return Str + Level; } }
